[#D3D-4573] Sets plugin version to 0.0.9
* Fixes Download api call for archive size:
"/download" endpoint is used to get download url and should be
used when the user actually wants to download the model.
It was requested when selecting a model in order to get
archive size, and it was incrementing model's downloadcount
(which should be done only if use clicks download)

This data has been moved to search results.

* Handles download url when it's outdated:
In the same session, if a user download a model several times during
the validity of the download url, download count is only incremented
once